Move refetch to one place
